# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: J-Desk API
  version: '2026-07-21'
  description: >
    Maschinenlesbare OpenAPI-3.1-Spezifikation der J-Desk-Server-API. Menschlich lesbare
    Referenz mit ausführlichen Beispielen, Fehlerbeschreibungen und Betriebsmodus-Hinweisen
    unter [docs/api/](README.md) (siehe insbesondere [rest.md](rest.md),
    [commands.md](commands.md), [websocket.md](websocket.md), [mcp.md](mcp.md)). Quelle für
    alle Angaben: `packages/server/src/app.ts`.


    Stand: Commit 4c47b3b — bei API-Änderungen mitpflegen.
servers:
  - url: http://localhost:4810
    description: Standard-Port (überschreibbar per PORT-Umgebungsvariable)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        `authorization: Bearer <TOKEN>`, ausgestellt von `POST /api/v1/auth/login` bzw.
        `POST /api/v1/auth/setup`.
  schemas:
    Error:
      type: object
      description: Einheitliches Fehlerformat aller Routen (deutscher Klartext).
      required: [error]
      properties:
        error:
          type: string
...
    DesktopState:
      type: object
      description: >
        Vollständige `DesktopState`-Struktur aus `@j-desk/core` (Docs, Stapel, Zettel,
        Verknüpfungen, Korb, Hintergrund, …). Absichtlich nicht feldweise modelliert — siehe
        `packages/core/src/model.ts` für die genaue Form; Feldbeispiele in commands.md.
        Provenienz: Docs, Links, Stapel, Zettel, Ausschnitte, Striche, Abdeckflächen, Stempel,
        Fahnen und Klammern tragen optional `createdBy` (string) und `createdAt` (string,
        ISO-8601-UTC) — vom Server gesetzt, nie vom Client; fehlt bei Bestandsobjekten.
        Papierkorb-Einträge (`trash[]`) tragen zusätzlich optional `trashedBy` (string). Details
        in commands.md → Provenienz. Fundstellen-Inhaltsmetadaten (kein Provenienz-Stempel,
        bei Kopien unverändert übernommen): Ausschnitte tragen zusätzlich optional
        `textSnapshot` (string, Client-geliefert, serverseitig auf 2000 Zeichen gekappt) und
        `fileSha256` (string, ausschließlich vom Server gesetzt, nur Standalone-Modus);
        Abdeckflächen tragen optional nur `textSnapshot`. Details in commands.md →
        Fundstellen-Provenienz. Objektversionen (dieselben zehn Arten, ohne trash/background):
        zusätzlich optional `updatedRev` (integer), `updatedAt` (string, ISO-8601-UTC) und
        `updatedBy` (string) — vom Server nach jedem Command zentral gestempelt, Grundlage der
        Konflikterkennung über `erwartet` bei `POST /desks/{id}/commands`. Details in
        commands.md → Objektversionen. jl-Referenzstatus (nur jlawyer-Modus, ausschließlich vom
        Server-Abgleich `GET /cases/{id}/desk` gepflegt): Docs tragen zusätzlich optional
        `sourceGone` (boolean, `true` wenn das Quelldokument in j-lawyer nicht mehr existiert
        — die Karte bleibt liegen, Annotationen unangetastet), `sourceChangeDate` (integer,
        Unix-Millis — jl-`changeDate` der Quelle zum Zeitpunkt des letzten Abgleichs) und
        `sourceReplacedAt` (string, ISO-8601-UTC — einmalig beim ersten Erkennen einer
        geänderten `changeDate` gesetzt, danach stabil bis zur nächsten Änderung). Details in
        rest.md → GET /cases/:id/desk.
    DeskStateResponse:
      type: object
      description: Antwortform aller Routen, die einen Schreibtisch-Zustand zurückgeben.
      required: [rev, state]
      properties:
        rev:
          type: integer
          description: Revisionszähler, erhöht sich bei jeder Änderung um genau eins.
        state:
          $ref: '#/components/schemas/DesktopState'
...
  responses:
    ErrorResponse:
      description: Fehlerantwort
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    DeskId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Schreibtisch-Id
...
paths:
...
  /api/v1/desks/{id}/state:
    get:
      operationId: getDeskState
      summary: Liest den aktuellen Zustand eines Schreibtischs
      tags: [Desks]
      parameters:
        - $ref: '#/components/parameters/DeskId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeskStateResponse'
        '404':
          description: Schreibtisch nicht gefunden
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: putDeskState
      summary: Ersetzt den gesamten Zustand eines Schreibtischs
      description: >
        Broadcastet das Ergebnis an alle verbundenen WebSocket-Clients desselben
        Schreibtischs. Für clientseitiges Undo über mehrere Schritte.
      tags: [Desks]
      parameters:
        - $ref: '#/components/parameters/DeskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DesktopState'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeskStateResponse'
        '400':
          description: Ungültiger Schreibtisch-Zustand (isValidState schlägt fehl)
          $ref: '#/components/responses/ErrorResponse'
        '404':
          description: Schreibtisch nicht gefunden
          $ref: '#/components/responses/ErrorResponse'
...